Updates player logging
Changes player projectile count to a debugGraph instead of a console log. This cleans up the log to be more readable.
Adds a logging prefix to the player class to make logs from it easy to identify.
Changes the Player Shooting log message to be a trace instead of a debug level log.